backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.0.0
python-dotenv>=1.0.0
langchain>=0.3.0
langchain-community>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=1.0.0
langchain-qdrant>=0.1.0
langchain-text-splitters>=0.0.1
google-generativeai>=0.5.0
qdrant-client>=1.9.0
pypdf>=4.2.0
pandas>=2.2.0
pyarrow>=16.0.0
scikit-learn>=1.5.0
tqdm>=4.66.0
httpx>=0.27.0
tenacity>=8.3.0
